requirements.txt

fastapi
uvicorn[standard]
python-telegram-bot==20.7
transformers
torch
python-multipart
